Refactor translations API and add documentation.
Reorganized and documented translation helper functions for improved readability and maintainability. Adjusted references to correct paths and updated key usage in `getCategoryTranslation`. Added JSDoc comments to clarify functionality.
